refactor: Add preset password to handle password hint and show/hide icon in build
