+++ QuickLaunchReactNativeKit/src/components/index.ts
@@ -2,7 +2,7 @@ export * from './Button';
 export * from './Icon';
 export * from './Screen';
 export * from './Text';
-export * from './TextField';
+export * from './TextField/TextField';
 export * from './Header';
 export * from './OTPTextField';
 export * from './Spinner';
